# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for managing tasks.",
        "title": "Task Management API",
        "contact": {
            "name": "Saran",
            "url": "github.com/saran-crayonte/",
            "email": "saran.kumaresan@crayonte.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v2/holiday": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Management"
                ],
                "summary": "Get all task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new holiday with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday Management"
                ],
                "summary": "Create a new holiday",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Holiday details",
                        "name": "holiday",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Holiday"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Holiday created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Holiday already defined",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v2/holiday/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a holiday by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday Management"
                ],
                "summary": "Get a holiday by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Holiday ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holiday retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Holiday"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Holiday not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing holiday by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday Management"
                ],
                "summary": "Update a holiday by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Updated holiday details",
                        "name": "holiday",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Holiday"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holiday updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Holiday not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an existing holiday by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday Management"
                ],
                "summary": "Delete a holiday by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Holiday ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holiday deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Holiday not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v2/task": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all holidays",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Holiday Management"
                ],
                "summary": "Get all holidays",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Holiday retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Holiday"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new task with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Management"
                ],
                "summary": "Create a new task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Task details",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Holiday": {
            "type": "object",
            "properties": {
                "holidayDate": {
                    "type": "string"
                },
                "holidayName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "properties": {
                "estimatedHours": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}